+++ src/codemagic/tools/app_store_connect.py
@@ -18,6 +18,8 @@ from codemagic.apple import AppStoreConnectApiError
 from codemagic.apple.app_store_connect import AppStoreConnectApiClient
 from codemagic.apple.app_store_connect import IssuerId
 from codemagic.apple.app_store_connect import KeyIdentifier
+from codemagic.apple.resources import Build
+from codemagic.apple.resources import BuildProcessingState
 from codemagic.apple.resources import BundleId
 from codemagic.apple.resources import BundleIdPlatform
 from codemagic.apple.resources import CertificateType
@@ -28,11 +30,14 @@ from codemagic.apple.resources import ProfileState
 from codemagic.apple.resources import ProfileType
 from codemagic.apple.resources import ResourceId
 from codemagic.apple.resources import SigningCertificate
+from codemagic.cli import Argument
 from codemagic.cli import Colors
 from codemagic.models import Certificate
 from codemagic.models import PrivateKey
 from codemagic.models import ProvisioningProfile
 from ._app_store_connect.arguments import AppStoreConnectArgument
+from ._app_store_connect.arguments import AppStoreVersionArgument
+from ._app_store_connect.arguments import BuildArgument
 from ._app_store_connect.arguments import BundleIdArgument
 from ._app_store_connect.arguments import CertificateArgument
 from ._app_store_connect.arguments import CommonArgument
@@ -62,7 +67,7 @@ def _get_certificate_key(
 class AppStoreConnect(cli.CliApp):
     """
     Utility to download code signing certificates and provisioning profiles
-    from Apple Developer Portal using App Store Connect API to perform iOS code signing.
+    from Apple Developer Portal using App Store Connect API to perform iOS code signing
     """
 
     def __init__(self,
@@ -148,6 +153,92 @@ class AppStoreConnect(cli.CliApp):
             else:
                 raise AppStoreConnectError(str(api_error))
 
+    @cli.action('list-builds',
+                BuildArgument.APPLICATION_ID_RESOURCE_ID_OPTIONAL,
+                BuildArgument.EXPIRED,
+                BuildArgument.NOT_EXPIRED,
+                BuildArgument.BUILD_ID_RESOURCE_ID,
+                BuildArgument.PRE_RELEASE_VERSION,
+                BuildArgument.PROCESSING_STATE,
+                BuildArgument.BUILD_VERSION_NUMBER)
+    def list_builds(self,
+                    application_id: Optional[ResourceId] = None,
+                    expired: Optional[bool] = None,
+                    not_expired: Optional[bool] = None,
+                    build_id: Optional[ResourceId] = None,
+                    pre_release_version: Optional[str] = None,
+                    processing_state: Optional[BuildProcessingState] = None,
+                    build_version_number: Optional[int] = None,
+                    should_print: bool = True) -> List[Build]:
+        """
+        List Builds from Apple Developer Portal matching given constraints
+        """
+        try:
+            expired_value = Argument.resolve_optional_two_way_switch(expired, not_expired)
+        except ValueError:
+            flags = f'{BuildArgument.EXPIRED.flags!r} and {BuildArgument.NOT_EXPIRED.flags!r}'
+            raise BuildArgument.NOT_EXPIRED.raise_argument_error(f'Using mutually exclusive switches {flags}.')
+
+        builds_filter = self.api_client.builds.Filter(
+            app=application_id,
+            expired=expired_value,
+            id=build_id,
+            processing_state=processing_state,
+            version=build_version_number,
+            pre_release_version_version=pre_release_version,
+        )
+        return self._list_resources(builds_filter, self.api_client.builds, should_print)
+
+    @classmethod
+    def _get_latest_build_number(cls, builds: List[Build]) -> Optional[int]:
+        try:
+            latest_build_number = max(int(build.attributes.version) for build in builds)
+        except ValueError:
+            return None
+        cls.echo(str(latest_build_number))
+        return latest_build_number
+
+    @cli.action('get-latest-app-store-build-number',
+                BuildArgument.APPLICATION_ID_RESOURCE_ID,
+                AppStoreVersionArgument.APP_STORE_VERSION)
+    def get_latest_app_store_build_number(self,
+                                          application_id: ResourceId,
+                                          app_store_version: Optional[str] = None,
+                                          should_print: bool = False) -> Optional[int]:
+        """
+        Get latest App Store build number for the given application
+        """
+        versions_client = self.api_client.app_store_versions
+        versions_filter = versions_client.Filter(version_string=app_store_version)
+        try:
+            _versions, builds = versions_client.list_with_include(
+                application_id, Build, resource_filter=versions_filter)
+        except AppStoreConnectApiError as api_error:
+            raise AppStoreConnectError(str(api_error))
+        self.printer.log_found(Build, builds, versions_filter)
+        self.printer.print_resources(builds, should_print)
+        return self._get_latest_build_number(builds)
+
+    @cli.action('get-latest-testflight-build-number',
+                BuildArgument.APPLICATION_ID_RESOURCE_ID,
+                BuildArgument.PRE_RELEASE_VERSION)
+    def get_latest_testflight_build_number(self,
+                                           application_id: ResourceId,
+                                           pre_release_version: Optional[str] = None,
+                                           should_print: bool = False) -> Optional[int]:
+        """
+        Get latest Testflight build number for the given application
+        """
+        versions_client = self.api_client.pre_release_versions
+        versions_filter = versions_client.Filter(app=application_id, version=pre_release_version)
+        try:
+            _versions, builds = versions_client.list_with_include(Build, resource_filter=versions_filter)
+        except AppStoreConnectApiError as api_error:
+            raise AppStoreConnectError(str(api_error))
+        self.printer.log_found(Build, builds, versions_filter)
+        self.printer.print_resources(builds, should_print)
+        return self._get_latest_build_number(builds)
+
     @cli.action('list-devices',
                 BundleIdArgument.PLATFORM_OPTIONAL,
                 DeviceArgument.DEVICE_NAME,
@@ -158,7 +249,7 @@ class AppStoreConnect(cli.CliApp):
                      device_status: Optional[DeviceStatus] = None,
                      should_print: bool = True) -> List[Device]:
         """
-        List Devices from Apple Developer portal matching given constraints.
+        List Devices from Apple Developer portal matching given constraints
         """
 
         device_filter = self.api_client.devices.Filter(
@@ -175,7 +266,7 @@ class AppStoreConnect(cli.CliApp):
                          platform: BundleIdPlatform = BundleIdPlatform.IOS,
                          should_print: bool = True) -> BundleId:
         """
-        Create Bundle ID in Apple Developer portal for specifier identifier.
+        Create Bundle ID in Apple Developer portal for specifier identifier
         """
 
         if bundle_id_name is None:
@@ -194,7 +285,7 @@ class AppStoreConnect(cli.CliApp):
                         platform: Optional[BundleIdPlatform] = None,
                         should_print: bool = True) -> List[BundleId]:
         """
-        List Bundle IDs from Apple Developer portal matching given constraints.
+        List Bundle IDs from Apple Developer portal matching given constraints
         """
 
         bundle_id_filter = self.api_client.bundle_ids.Filter(
@@ -208,7 +299,7 @@ class AppStoreConnect(cli.CliApp):
                       bundle_id_resource_id: ResourceId,
                       should_print: bool = True) -> BundleId:
         """
-        Get specified Bundle ID from Apple Developer portal.
+        Get specified Bundle ID from Apple Developer portal
         """
 
         return self._get_resource(bundle_id_resource_id, self.api_client.bundle_ids, should_print)
@@ -220,7 +311,7 @@ class AppStoreConnect(cli.CliApp):
                          bundle_id_resource_id: ResourceId,
                          ignore_not_found: bool = False) -> None:
         """
-        Delete specified Bundle ID from Apple Developer portal.
+        Delete specified Bundle ID from Apple Developer portal
         """
 
         self._delete_resource(self.api_client.bundle_ids, bundle_id_resource_id, ignore_not_found)
@@ -270,7 +361,7 @@ class AppStoreConnect(cli.CliApp):
                         save: bool = False,
                         should_print: bool = True) -> SigningCertificate:
         """
-        Get specified Signing Certificate from Apple Developer portal.
+        Get specified Signing Certificate from Apple Developer portal
         """
 
         private_key = _get_certificate_key(certificate_key, certificate_key_password)
@@ -292,7 +383,7 @@ class AppStoreConnect(cli.CliApp):
                            certificate_resource_id: ResourceId,
                            ignore_not_found: bool = False) -> None:
         """
-        Delete specified Signing Certificate from Apple Developer portal.
+        Delete specified Signing Certificate from Apple Developer portal
         """
 
         self._delete_resource(self.api_client.signing_certificates, certificate_resource_id, ignore_not_found)
@@ -315,7 +406,7 @@ class AppStoreConnect(cli.CliApp):
                           save: bool = False,
                           should_print: bool = True) -> List[SigningCertificate]:
         """
-        List Signing Certificates from Apple Developer Portal matching given constraints.
+        List Signing Certificates from Apple Developer Portal matching given constraints
         """
 
         private_key = _get_certificate_key(certificate_key, certificate_key_password)
@@ -389,7 +480,7 @@ class AppStoreConnect(cli.CliApp):
     @cli.action('get-profile', ProfileArgument.PROFILE_RESOURCE_ID)
     def get_profile(self, profile_resource_id: ResourceId, should_print: bool = True) -> Profile:
         """
-        Get specified Profile from Apple Developer portal.
+        Get specified Profile from Apple Developer portal
         """
 
         return self._get_resource(profile_resource_id, self.api_client.profiles, should_print)
@@ -401,7 +492,7 @@ class AppStoreConnect(cli.CliApp):
                        profile_resource_id: ResourceId,
                        ignore_not_found: bool = False) -> None:
         """
-        Delete specified Profile from Apple Developer portal.
+        Delete specified Profile from Apple Developer portal
         """
 
         self._delete_resource(self.api_client.profiles, profile_resource_id, ignore_not_found)
@@ -418,7 +509,7 @@ class AppStoreConnect(cli.CliApp):
                       save: bool = False,
                       should_print: bool = True) -> List[Profile]:
         """
-        List Profiles from Apple Developer portal matching given constraints.
+        List Profiles from Apple Developer portal matching given constraints
         """
         profile_filter = self.api_client.profiles.Filter(
             profile_type=profile_type,
@@ -455,7 +546,7 @@ class AppStoreConnect(cli.CliApp):
                                 save: bool = False,
                                 should_print: bool = True) -> List[Profile]:
         """
-        List provisioning profiles from Apple Developer Portal for specified Bundle IDs.
+        List provisioning profiles from Apple Developer Portal for specified Bundle IDs
         """
 
         profiles_filter = self.api_client.profiles.Filter(
@@ -490,7 +581,7 @@ class AppStoreConnect(cli.CliApp):
                             create_resource: bool = False) -> Tuple[List[Profile], List[SigningCertificate]]:
         """
         Fetch provisioning profiles and code signing certificates
-        for Bundle ID with given identifier.
+        for Bundle ID with given identifier
         """
 
         private_key = _get_certificate_key(certificate_key, certificate_key_password)