+++ tests/apple/app_store_connect/builds/test_builds.py
@@ -1,0 +1,21 @@
+import os
+
+import pytest
+
+from codemagic.apple.resources import Build
+from codemagic.apple.resources import ResourceType
+from tests.apple.app_store_connect.resource_manager_test_base import ResourceManagerTestsBase
+
+
+@pytest.mark.skipif(not os.environ.get('RUN_LIVE_API_TESTS'), reason='Live App Store Connect API access')
+class BuildsTest(ResourceManagerTestsBase):
+    def test_list(self):
+        builds = self.api_client.builds.list()
+        assert len(builds) > 0
+        for build in builds:
+            assert isinstance(build, Build)
+            assert build.type is ResourceType.BUILDS
+
+    def test_filter(self):
+        assert self.api_client.builds.Filter._get_field_name(
+            'pre_release_version_version') == 'preReleaseVersion.version'