+++ tests/apple/resources/test_resource.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from datetime import timedelta
 from datetime import timezone
 
 import pytest
@@ -22,6 +23,7 @@ def test_to_iso_8601(iso_8601_timestamp, expected_datetime):
 @pytest.mark.parametrize('given_datetime, expected_iso_8601_timestamp', [
     (None, None),
     ('2020-08-04T11:44:12.000+0000', datetime(2020, 8, 4, 11, 44, 12, tzinfo=timezone.utc)),
+    ('2021-01-28T06:01:32-08:00', datetime(2021, 1, 28, 6, 1, 32, tzinfo=timezone(timedelta(days=-1, seconds=57600)))),
     ('1970-01-01T00:00:00.000+0000', datetime(1970, 1, 1, 0, 0, 0, tzinfo=timezone.utc)),
 ])
 def test_from_iso_8601(given_datetime, expected_iso_8601_timestamp):